[sshd] Adapt default path
Previously, sshd used a default path, which used the PATH of a regular
Unix distro.

$> ssh server 'echo $PATH'
/usr/bin:/bin:/usr/sbin:/sbin:/<nix-store-openssh>/bin

So, when you had a non-interactive shell, it could have been the case,
that /etc/profile did not get sourced, so a NixOS-incompatible PATH was
in the environment.

This patch adds a default PATH in OpenSSH to match the NixOS
configuration by default. So shells, which do not source /etc/profile in
their invocation, will also have the ability to find and execute
commands.

An example shell, which does not source /etc/profile in non-interactive
mode is zsh or (plain old) sh.

Note: This patch only adds system PATHs to the environment, as OpenSSH
does not seem to evaluate variables of a tilde.